feat!: publicKey to verificationMethod as correction
BREAKING CHANGE: publicKey is no longer accepted by default. To achieve
the same behavior, the document must have the correction created with
this commit applied first. This can be done by calling it directly with
the dictionary value of the document or by using `deserialize_document`
with the correction method specified in the corrections list. See docs
on pydid.doc.corrections.public_key_is_verification_method for examples.

Signed-off-by: Daniel Bluhm <dbluhm@pm.me>