feat: realign service with DID Core spec
This PR realigns the service endpoint definitions to what is outlined in
the DID Core spec. Specifically, permitting the service endpoint to be a
string, map, or set of strings and/or maps. DIDs, DID URLs, general
URLs, and now any string is now permitted. This is to support specifying
arbitrary URIs. The rules for parsing URIs would be overly complicated
to fully validate; slightly greater flexibility is preferrable to
maintaining large regexes.

Signed-off-by: Daniel Bluhm <dbluhm@pm.me>